:recycle: Extract DatastorePackageFactory from DatastoreSupported's singleton
