+++ lib/goon_model_gen/builder/store_builder.rb
@@ -5,7 +5,7 @@ require "goon_model_gen/builder/abstract_builder"
 require "goon_model_gen/source/struct"
 
 require "goon_model_gen/golang/package"
-require "goon_model_gen/golang/datastore_supported"
+require "goon_model_gen/golang/datastore_package_factory"
 
 
 module GoonModelGen
@@ -44,7 +44,7 @@ module GoonModelGen
 
       # @param pkgs [Golang::Packages]
       def resolve_type_names(pkgs)
-        pkgs.resolve_type_names(Golang::DatastoreSupported.packages.dup.add(model_packages))
+        pkgs.resolve_type_names(Golang::DatastorePackageFactory.new.packages.dup.add(model_packages))
       end
 
     end
