+++ lib/goon_model_gen/builder/validation_builder.rb
@@ -3,7 +3,7 @@ require "goon_model_gen"
 require "goon_model_gen/builder/abstract_builder"
 
 require "goon_model_gen/golang/packages"
-require "goon_model_gen/golang/datastore_supported"
+require "goon_model_gen/golang/datastore_package_factory"
 
 
 module GoonModelGen
@@ -24,7 +24,7 @@ module GoonModelGen
 
       # @param pkgs [Golang::Packages]
       def resolve_type_names(pkgs)
-        pkgs.resolve_type_names(Golang::DatastoreSupported.packages)
+        pkgs.resolve_type_names(Golang::DatastorePackageFactory.new.packages)
       end
 
     end
